frameworks: omap-enhancement: multi-gpu egl support
In order to support several GPU types from the same filesystem we
have expanded the egl.cfg definition to optionally support a cpu type.

The format of an egl.cfg entry is as follows:

display-id index egl-library-string [cpu]

where [cpu] is determined from the Android property ro.board.processor
(which is derived from the /sys/board_properties/soc/family)

When ro.product.processor is set the EGL library will verify the cpu
type when selecting an EGL implementation. When not set a cpu does not
need to be specified.

Change-Id: I3b8b758af643f96f3e624c9de8a970660c84b1f0
Signed-off-by: Daniel Levin <dendy@ti.com>